Add field to get URLs in a new tab
Some email sending could fail from mailgun. So we would like to be able to send manual
invitations.

Let's add a field in the admin panel where the admin can enter a number of URLs.

The URLs will appear in a new browser tab in the same way as generated QR codes appear
in a new tab, except they will be in text format, not as QR codes.