Fixed: show error when election is closed
When user is on /votes and election has been close and user try to send their
vote, then there was the usual animation and then nothing. An error was printed
on the console but the user was not informed on the web page that the election
has been closed and that their vote wasn't taken into account.

Now we properly display an error message in that case.